# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Wedding guest-list backend. Collaborative RSVP management with\nrole-based access (owner/editor/viewer), invite links, stats,\nactivity feed, notifications, and CSV export.\n\nAll responses use the envelope `{ \"data\": ..., \"error\": null }`.\nOn error, `data` is null and `error` is `{ \"code\", \"message\", \"details\" }`.\nAuthenticate via `POST /auth/login`, then send `Authorization: Bearer \u003caccess\u003e`.",
        "title": "Zawaj API",
        "contact": {
            "name": "Zawaj"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/invite/{token}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns wedding name and role for a share token before accepting.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invites"
                ],
                "summary": "Preview an invite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invite token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/zawaj_pkg_response.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/invite/{token}/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Joins the wedding at the link's role; never downgrades an existing higher role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invites"
                ],
                "summary": "Accept an invite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invite token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/zawaj_pkg_response.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/zawaj_pkg_response.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/weddings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weddings"
                ],
                "summary": "List my weddings",
                "responses": {
                    "200": {
                        "description": "{ items: [...] }",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/zawaj_pkg_response.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weddings"
                ],
                "summary": "Create a wedding",
                "parameters": [
                    {
                        "description": "Wedding",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/zawaj_internal_dto.CreateWeddingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/zawaj_pkg_response.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/zawaj_pkg_response.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/weddings/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weddings"
                ],
                "summary": "Get a wedding",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wedding ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ wedding, my_role }",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/zawaj_pkg_response.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/zawaj_pkg_response.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/zawaj_pkg_response.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weddings"
                ],
                "summary": "Delete a wedding (owner only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wedding ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ deleted: true }",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/zawaj_pkg_response.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weddings"
                ],
                "summary": "Update a wedding (owner only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wedding ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/zawaj_internal_dto.UpdateWeddingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/zawaj_pkg_response.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/zawaj_pkg_response.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/weddings/{id}/members": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "members"
                ],
                "summary": "List members",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wedding ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ items: [...] }",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/zawaj_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/zawaj_pkg_response.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "zawaj_internal_dto.CreateWeddingRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 2000
                },
                "event_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 120
                }
            }
        },
...
        "zawaj_internal_dto.UpdateWeddingRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 2000
                },
                "event_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 120
                }
            }
        },
...
        "zawaj_pkg_response.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "zawaj_pkg_response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/zawaj_pkg_response.APIError"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the access JWT.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}